Move SuperSimplex to Core
