+++ src/core.rs
@@ -2,4 +2,5 @@ pub mod open_simplex;
 pub mod perlin;
 pub mod perlin_surflet;
 pub mod simplex;
+pub mod super_simplex;
 pub mod worley;
